requirements.txt
Flask==2.2.2
keras==2.11.0
python-dotenv==0.21.1
tensorflow==2.11.0
